[#2] Refactor : 불필요한 코드 제거 및 정리
