+++ src/index.tsx
@@ -1,5 +1,9 @@
 import ReactDOM from 'react-dom/client';
-import { RouterProvider, createBrowserRouter } from 'react-router-dom';
+import {
+  Navigate,
+  RouterProvider,
+  createBrowserRouter,
+} from 'react-router-dom';
 
 import App from './App';
 import './index.css';
@@ -9,12 +13,15 @@ const router = createBrowserRouter([
   {
     path: '/',
     element: <App />,
-    errorElement: <div />,
     children: [
       {
         index: true,
-        path: '/home',
+        path: '/',
         element: <MainPage />,
+      },
+      {
+        path: '*',
+        element: <Navigate replace to='/' />,
       },
     ],
   },
